requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==0.8.1
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
defusedxml==0.6.0
FFC==2017.1.0
FIAT==2017.1.0
gast==0.3.3
google-pasta==0.2.0
graphviz==0.14
grpcio==1.28.1
h5py==2.10.0
idna==2.9
image-classifiers==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.5.0
networkx==2.2
numpy==1.15.4
olefile==0.46
parso==0.4.0
Pillow==7.1.1
protobuf==3.11.3
pydicom==1.2.2
pydot==1.4.1
Pygments==2.4.2
pyparsing==2.3.1
python-dateutil==2.8.0
PyYAML==5.3.1
qtconsole==4.5.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
sympy==1.1.1
tb-nightly==1.14.0a20190603
TBB==0.1
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tf2cv==0.0.11
toolz==0.9.0
torch==1.4.0
torchvision==0.5.0
UFL==2017.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
